streamlit_cetes_app/requirements.txt
streamlit
pandas
numpy
requests
Pillow
matplotlib
